+++ CHANGELOG.md
@@ -1,7 +1,11 @@
-## [0.0.1] - 24 Sept 2020
+## [0.0.2] - 16 May 2021
 
-* A working class Colours with more swatches and function swatch to generate customs swatch
+* Migrate to null safety
 
 ## [0.0.1+1] - 24 Sept 2020
 
-* Update README.md
+* Update README
+
+## [0.0.1] - 24 Sept 2020
+
+* A working class Colours with more swatches and function swatch to generate customs swatch